+++ benchmark/benchmarks.jl
@@ -12,3 +12,4 @@ end
 const SUITE = BenchmarkGroup()
 
 @include_bm SUITE "bm_assembly"
+@include_bm SUITE "bm_monomial_basis"