+++ src/Adaptivity/Adaptivity.jl
@@ -40,6 +40,8 @@ export AdaptedTriangulation
 export Triangulation, is_change_possible, best_target, get_adapted_model
 export change_domain, move_contributions
 
+export DorflerMarking, mark, estimate
+
 include("RefinementRules.jl")
 include("FineToCoarseFields.jl")
 include("OldToNewFields.jl")
@@ -51,5 +53,6 @@ include("MacroFEs.jl")
 include("CompositeQuadratures.jl")
 include("EdgeBasedRefinement.jl")
 include("SimplexifyRefinement.jl")
+include("AdaptiveMeshRefinement.jl")
 
 end # module